mobicore: initial import of kernel driver
MobiCore kernel driver is an interface to the mobicore TEE
This version uses under_score coding style only

MobiCore is an operating system being shipped with TZBSP
on msm chipsets. MobiCore consists of several components in
the secure world(TrustZone) and non-secure world(linux
kernel, Android user space). The MobiCore driver
communicates with the MobiCore kernel that exists in
TrustZone.

Signed-off-by: Lukas Hänel <lukas.haenel@gi-de.com>